backend: Restructure project into a modular architecture
