--- backend/.env
@@ -1,3 +1,0 @@
-PORT=5000
-DATABASE_PATH=database/temperature.db
-DEBUG=True
