+++ backend/services/weather_fetcher.py
@@ -1,16 +1,15 @@
+import sqlite3
 import requests
 import time
+import traceback
 from datetime import datetime
-from models import get_db_connection, DEFAULT_LATITUDE, DEFAULT_LONGITUDE
-from app import predict_for_day
+from db.connection import get_db_connection
+from db.maintenance import DEFAULT_LATITUDE, DEFAULT_LONGITUDE
 
 def get_current_temperature():
     """Get current temperature from Open-Meteo Forecast API and store it in database"""
     try:
-        # Open-Meteo Forecast API endpoint
         url = "https://api.open-meteo.com/v1/forecast"
-        
-        # Parameters to get current weather for Agadir
         params = {
             "latitude": DEFAULT_LATITUDE,
             "longitude": DEFAULT_LONGITUDE,
@@ -19,7 +18,6 @@ def get_current_temperature():
             "timezone": "auto"
         }
         
-        # Make GET request
         response = requests.get(url, params=params)
         
         if response.ok:
@@ -29,7 +27,6 @@ def get_current_temperature():
                 current_temp = data["current_weather"]["temperature"]
                 timestamp = datetime.now().isoformat()
                 
-                # Store in database
                 conn = get_db_connection()
                 cursor = conn.cursor()
                 
@@ -42,7 +39,6 @@ def get_current_temperature():
                     conn.commit()
                     print(f"[{timestamp}] Temperature stored: {current_temp:.2f}°C")
                     
-                    # Get the last 10 readings for this hour
                     cursor.execute('''
                     SELECT temperature 
                     FROM temperature_data 
@@ -72,41 +68,5 @@ def get_current_temperature():
             
     except Exception as e:
         print(f"Error getting current temperature: {str(e)}")
-        import traceback
         traceback.print_exc()
         raise
-
-def update_all_predictions():
-    """
-    Update all predictions for the next 5 days.
-    This completely refreshes the predictions table daily.
-    """
-    try:
-        print(f"[{datetime.now().isoformat()}] Starting daily prediction update for next 5 days...")
-        
-        conn = get_db_connection()
-        cursor = conn.cursor()        
-        cursor.execute('DELETE FROM temperature_predictions')
-        conn.commit()
-        print("Cleared existing predictions")
-        
-        prediction_count = 0
-        for day in range(1, 6):
-            try:
-                result = predict_for_day(day)
-                if "error" in result:
-                    print(f"Error predicting day {day}: {result['error']}")
-                else:
-                    prediction_count += len(result.get("predictions", []))
-                    print(f"Successfully generated predictions for day {day}")
-            except Exception as e:
-                print(f"Error processing day {day}: {str(e)}")
-                continue
-        
-        print(f"[{datetime.now().isoformat()}] Successfully generated {prediction_count} hourly predictions for next 5 days")
-        return True
-    except Exception as e:
-        print(f"Error updating predictions: {str(e)}")
-        import traceback
-        traceback.print_exc()
-        return False
